refactor(reporter): carve data::Report + split formats into siblings (#319)
The 2028-line reporter/mod.rs held the canonical JsonReport shape,
its 13 sub-structs, from_audit, build_json_dir_tree, two path
helpers, three formats (xml, sonar, pr), plus the test suite.

Now:
- reporter/data.rs (562 LOC) owns the canonical Report shape +
  builder + dir-tree walker + path helpers.
- reporter/{xml,sonar,text,pr,briefing}.rs each own their format.
- reporter/mod.rs is a thin re-export shell — ~50 LOC of
  production code plus the existing test suite. mod.rs file size
  is 639 because the cfg(test) block + _format_markdown_single
  helper stay put; the production surface fits in 50.

External import paths unchanged: reporter::JsonReport,
format_xml/sonar/text/monorepo_text/pr/briefing all resolve via
re-exports in mod.rs. No behaviour change.

Pairs with #317 (widen ReportFormatter): once Report is a real
module, the bespoke arms for md/html/bundle/dashboard/pr/
explorer/strategy in commands/report.rs become mechanical
adapter migrations.

All 308 tests green; clippy -D warnings clean; noupling's own
audit unchanged at 98.3/100.
