+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- **Reporter: carve `data::Report` (`JsonReport`) into its own module + split each format out** (#319): `reporter/mod.rs` used to hold the canonical `JsonReport` shape, its 13 sub-structs, the `from_audit` builder, the `build_json_dir_tree` walker, two path helpers, *and* three formats (`format_xml`, `format_sonar`, `format_pr`) *and* the test suite — 2028 LOC. Issue #319 carves the wire shape into `reporter/data.rs` (562 LOC, the canonical `Report`), and splits each format into its own sibling: `xml.rs` (142 LOC), `sonar.rs` (124 LOC), `text.rs` (382 LOC — owns `format_text` and `format_monorepo_text`), `pr.rs` (117 LOC), `briefing.rs` (136 LOC). `mod.rs` becomes a thin re-export shell — ~50 LOC of production code plus the pre-existing test suite. No external import paths change: `reporter::JsonReport`, `reporter::format_xml`, `reporter::format_sonar`, `reporter::format_text`, `reporter::format_monorepo_text`, `reporter::format_pr`, `reporter::format_briefing` all keep working through unchanged re-exports. Pairs with #317 (widen the `ReportFormatter` seam): once the canonical `Report` shape is its own module, folding md/html/bundle/dashboard/explorer/strategy adapters into the same registry as xml/sonar/text/pr/briefing becomes a mechanical follow-up. All 308 tests green; clippy `-D warnings` green; noupling's own audit unchanged at 98.3.
+
 - **Explorer: extract `IssueFocus` module from `App.tsx`** (#316): The canvas-level focus state driven by Issues-tab selections — `participantsOf`, `longestCommonAncestor`, expanded-container derivation, and the inter-participant edge set — moves into `crates/noupling-explorer/template/src/state/issueFocus.ts` behind one public function `computeIssueFocus(issue, key, data) → IssueFocus`. `App.tsx`'s `onIssueFocus` handler shrinks from ~50 inline lines to a four-line call. Pure function, no React, no setState; callers still own the side effect of scoping the view to `focus.lca`. Unit-test framework is not yet wired in the template subproject (Playwright e2e only), so verification is the existing smoke suite — all 30 specs still pass. Lays groundwork for #318 (HighlightPolicy) where the focus rule becomes one input to a unified canvas-highlight policy instead of a tangled sibling of `pathFinder` and `minCut` memos.
 
 - **Workspace split: `noupling-core` / `noupling` (cli) / `noupling-explorer`** (#229, Task 1 of #228): The single-crate repo became a Cargo workspace with three crates under `crates/`. `noupling-core` is the pure analysis library (analyzer, scanner, storage, layers/rules, settings, baseline, diff). The cli crate (package name still `noupling`, binary still `noupling`) is the only consumer that wires reporter + cli code; it depends on `noupling-core` and `noupling-explorer`. `noupling-explorer` is an empty stub crate ready for Task 2 of #228. Dependency direction is strictly one-way: `cli → explorer → core`, `cli → core`. The boundary is dogfooded in `.noupling/settings.json` via new `dependency_rules` that forbid core → cli/explorer and explorer → cli/reporter; the same boundary is enforced at compile time by the workspace layout. The `analyzer::AuditResultBuilder` test helper is now exposed via a `test-utils` cargo feature on `noupling-core`, enabled as a dev-dependency feature on the cli crate. CI (`cargo check/clippy/test`) switched to `--workspace` and `cargo fmt` switched to `--all`. `scripts/release.sh` likewise. Binary surface is unchanged (`noupling --help` byte-identical, all integration tests pass without modification). **Coordinated follow-ups required before the next release tag:** Homebrew tap formula must be updated to `cargo install --path crates/noupling-cli --bin noupling`; crates.io publish flow needs an explicit decision (publish core/explorer/cli in dep order, or drop crates.io publishing) — see #229 for the checklist.
